requirements/prod.txt
psycopg==2.7.1
django-storages==1.1.8
